requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.0
cffi==1.7.0
cloudinary==1.5.0
decorator==4.0.10
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
django-allauth==0.25.2
django-allauth-bootstrap==0.2
django-anymail==0.6
django-booking==0.7.2
django-bootstrap-form==3.2.1
django-countries==4.0
django-font-awesome==4.6.1
django-hvad==1.6.0
django-libs==1.67.7
django-mailgun==0.9.1
django-toolbelt==0.0.1
enum34==1.1.6
envvars==0.3.0
gunicorn==19.6.0
hashids==1.1.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
oauthlib==1.1.2
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.1
prompt-toolkit==1.0.8
psycopg2==2.6.1
ptyprocess==0.5.1
pycparser==2.14
Pygments==2.1.3
python-openid==2.2.5
requests==2.11.1
requests-oauthlib==0.6.2
sendgrid==1.5.14
simplegeneric==0.8.1
six==1.10.0
smtpapi==0.2.0
static3==0.7.0
traitlets==4.3.1
urllib3==1.19.1
wcwidth==0.1.7
wheel==0.24.0
whitenoise==3.2.2
